Fixed a major bug that was causing the map optimisation to sometimes oscillate, causing residuals to grow. Fixed a bug with map warping using last observed to use scan correction not overall error. Fixed a few bugs with map warping by every scan

+++ crosbot_graphslam/include/crosbot_graphslam/graphSlamCPU.hpp
@@ -85,7 +85,7 @@ private:
       double warpPointsY[MAX_LOCAL_POINTS];
       double warpPointsZ[MAX_LOCAL_POINTS];
       int lastObserved[MAX_LOCAL_POINTS];
-      double numWarpPoints;
+      int numWarpPoints;
 
       vector<Scan *> scans;
    } LocalMap;
